fix: implemented next appointment as a mixin
